COM-3988: create dashboard
