+++ src/core/composables/courseBills.js
@@ -1,18 +1,25 @@
 import { ref } from 'vue';
 import { useQuasar } from 'quasar';
 import get from 'lodash/get';
+import Companies from '@api/Companies';
 import CourseBills from '@api/CourseBills';
+import CourseBillingItems from '@api/CourseBillingItems';
+import CourseFundingOrganisations from '@api/CourseFundingOrganisations';
 import CourseCreditNotes from '@api/CourseCreditNotes';
 import { NotifyNegative, NotifyPositive } from '@components/popup/notify';
-import { formatDownloadName, formatQuantity } from '@helpers/utils';
+import { formatDownloadName, formatQuantity, formatAndSortCompanyOptions, formatAndSortOptions } from '@helpers/utils';
 import { downloadFile } from '@helpers/file';
-import { REQUIRED_LABEL } from '../data/constants';
+import { descendingSortBy } from '@helpers/dates/utils';
+import { COMPANY, REQUIRED_LABEL, FUNDING_ORGANISATION, DIRECTORY } from '../data/constants';
 
 export const useCourseBilling = (courseBills, validations, refreshCourseBills) => {
   const $q = useQuasar();
 
   const pdfLoading = ref(false);
   const selectedBills = ref([]);
+  const payerList = ref([]);
+  const billingItemList = ref([]);
+  const areDetailsVisible = ref(Object.fromEntries(courseBills.value.map(bill => [bill._id, false])));
 
   const downloadBill = async (bill) => {
     try {
@@ -103,15 +110,55 @@ export const useCourseBilling = (courseBills, validations, refreshCourseBills) =
       .onCancel(() => NotifyPositive('Suppression annulée.'));
   };
 
+  const refreshPayers = async () => {
+    try {
+      const organisations = await CourseFundingOrganisations.list();
+      const companyList = await Companies.list({ action: DIRECTORY });
+      const formattedOrganisationList = formatAndSortOptions(organisations, 'name');
+      const formattedCompanyList = formatAndSortCompanyOptions(companyList, 'name');
+      payerList.value =
+          [
+            ...formattedOrganisationList.map(payer => ({ ...payer, type: FUNDING_ORGANISATION })),
+            ...formattedCompanyList.map(company => ({ ...company, type: COMPANY })),
+          ];
+    } catch (e) {
+      console.error(e);
+      payerList.value = [];
+      NotifyNegative('Erreur lors de la récupération des financeurs.');
+    }
+  };
+
+  const refreshBillingItems = async () => {
+    try {
+      const billingItems = await CourseBillingItems.list();
+      billingItemList.value = formatAndSortOptions([...billingItems], 'name');
+    } catch (e) {
+      console.error(e);
+      billingItemList.value = [];
+      NotifyNegative('Erreur lors de la récupération des articles de facturation.');
+    }
+  };
+
+  const unrollBill = (billId) => {
+    const bill = billId || [...courseBills.value].sort(descendingSortBy('createdAt'))[0]._id;
+    areDetailsVisible.value[bill] = !areDetailsVisible.value[bill];
+  };
+
   return {
     // Data
     pdfLoading,
     selectedBills,
+    payerList,
+    billingItemList,
+    areDetailsVisible,
     // Methods
     downloadBill,
     downloadCreditNote,
     getBillErrorMessages,
     updateSelectedBills,
     openBillDeletionModal,
+    refreshPayers,
+    refreshBillingItems,
+    unrollBill,
   };
 };
